+++ src/app/PubSubVitessceGrid.js
@@ -2,7 +2,6 @@ import React from 'react';
 import VitessceGrid from 'vitessce-grid';
 
 import { LayerPublisher } from '../components/layerpublisher';
-import { getPreferredTheme } from '../components/utils';
 
 export default class PubSubVitessceGrid extends React.Component {
   constructor(props) {
@@ -12,9 +11,8 @@ export default class PubSubVitessceGrid extends React.Component {
   }
 
   render() {
-    const { config, getComponent } = this.props;
+    const { config, getComponent, theme } = this.props;
     const { allReady } = this.state;
-    const theme = config.theme || getPreferredTheme();
     return (
       <div className={`vitessce-container vitessce-theme-${theme}`}>
         { allReady && <LayerPublisher layers={config.layers} /> }